Add GPS command and handlers to the Quectel BG9X modem driver
Remove the modem init retry logic and add a work queue for the PDP
context activation. This will fix the pend boot issue where the modem would
not register on the network after a reboot.

And add a work queue for the GPS query command.
This code is not yet implemented in the modem driver.